Refactored schema

+++ action/medium/delete.go
@@ -55,6 +55,17 @@ func delete(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
+	// check if medium is associated with charts
+	var totAssociated int
+	config.DB.Model(&model.Chart{}).Where(&model.Chart{
+		FeaturedMediumID: uint(id),
+	}).Count(&totAssociated)
+
+	if totAssociated != 0 {
+		errorx.Render(w, errorx.Parser(errorx.CannotSaveChanges()))
+		return
+	}
+
 	config.DB.Delete(&result)
 
 	renderx.JSON(w, http.StatusOK, nil)
